Grouped photo to pdf functions
