+++ pdf_bot/constants.py
@@ -12,7 +12,7 @@ PDF_INVALID_FORMAT = 1
 PDF_TOO_LARGE = 2
 
 # PDF file constants
-WAIT_TASK = 0
+WAIT_DOC_TASK = 0
 WAIT_DECRYPT_PW = 1
 WAIT_ENCRYPT_PW = 2
 WAIT_ROTATE_DEGREE = 3
@@ -27,6 +27,7 @@ WAIT_CROP_PERCENT = 11
 WAIT_CROP_OFFSET = 12
 WAIT_EXTRACT_PHOTO_TYPE = 13
 WAIT_TO_PHOTO_TYPE = 14
+WAIT_PHOTO_TASK = 15
 
 # Keyboard constants
 CANCEL = _('Cancel')
